Refactor calendar tests
